更新代码

+++ app/model/w_channel.go
@@ -5,18 +5,18 @@ import (
 )
 
 type WChannel struct {
-	Id          int    `gorm:"primaryKey;autoIncrement"`
-	Title       string `json:"title"`
-	Status      int    `json:"status"`
-	Max         string `json:"max" gorm:"default:99999"`
-	Min         string `json:"min" gorm:"default:0"`
-	DefaultRate int    `json:"defaultRate"`
-	Coding      string `json:"coding"`
-	AddRate     int    `json:"addRate"`
-	Type        int    `json:"type"`
-	Charge      int    `json:"charge"`
-	Currency    string `json:"currency"`
-	Remark      string `json:"remark"`
+	Id          int     `gorm:"primaryKey;autoIncrement"`
+	Title       string  `json:"title"`
+	Status      int     `json:"status"`
+	Max         string  `json:"max" gorm:"default:99999"`
+	Min         string  `json:"min" gorm:"default:0"`
+	DefaultRate float64 `json:"defaultRate"`
+	Coding      string  `json:"coding"`
+	AddRate     float64 `json:"addRate"`
+	Type        int     `json:"type"`
+	Charge      int     `json:"charge"`
+	Currency    string  `json:"currency"`
+	Remark      string  `json:"remark"`
 	CreateBy    string
 	CreateTime  datetime.Datetime `gorm:"autoCreateTime"`
 	UpdateBy    string
